This version of the monitor succesfully handles the situation where timing out based on time doesn't always work.
It has been observed on Linux and SUN systems that specifying a
sleep/select time of 1 second will wait 0.99965 seconds.  When comparing
this against 1 it is less than a therefore things were not timeing out.
This version handles this case correctly on Linux and SUN systems.  It
should be noted that IRIX machines worked fine without this changed becuase
those systems waited the full time.