Fixed exit status bug.  Now it will return 1 for error, 0 for success instead of 0 every time.
Fixed output of data access layer to actually output usefull info to the
fields instead of blanks.

Removed some newlines from text sent to the log file.  The newlines are
there for readability on the terminal and can screew up log file parsers
that look for one entry per line.