Handles 'retries' and 'resubmitions' in their own appropriate way.
RESUBMITTING is displayed instead of NET_ERROR because this confused people.

NOACCESS errors are checked durring the transfer setup process, not just at
the very beginning (on reads).

The control socket is checked for messages just before encp would block
trying to write.  Most likely this is from a mount failure.