Added two columns to the VOLUME_QUOTAS page. 1) Requested # of tapes 2) Authorized # of tapes
The file where this info is stored is specified in the 'inventory' section
of the config file under 'authorized_file'.  This file can contain comments
that begin on a line with #.  Comment lines and empty lines are ignored.
Normal lines contain four items.  The first two are a (library,
storage_group) tuple.  The last two are integers coresponding to #1 and
#2 above.  Lastly, there can be one line of the form
blanks=X
where X is an integer corresponding to the number of blanks tapes in the
robot.