Discipline functionality implemented. It is to restrict acceptiong requsts from a certain Storage Group based on the arbitrary criteria. The configuration dctionary has to contain a 'discipline' key. Examle is below: configdict['discipline'] = {'D0':{1:{'keys':{'host':"happy+"},                                      'function':'restrict_host_access',                                      'args':['D0','happy.fnal.gov',1],                                      'action':'ignore'},                                   2:{'keys':{'version': 'v+'},                                      'function':'restrict_version_access',                                      'args':['D0','v2_15',{}],                                      'action':'reject'},
                               },
                           }
Here we see a one SG to be restricted. It has two choices. They are processed
in the order from higher to lower: 2 then 1. 'keys' are keys that will be
matched with the in coming request ticket keys. If match occurs 'function'
will be called with arguments 'args' if fuction returns 'true' (not 0), then
action will be performed with the incoming request.