The checkbutton on the menubar is now set to on initially.
Also, the resize event -- from a user -- waits a finite amount of time until
the window contents are redrawn.  Should extra resize events occur the wait time
is reset to the full wait time.  This should cut down on the number of times
the canvas is redrawn when the user is resizing the window.