This is the first version of enstore's ensync capability.  It works similar to rsync.
It takes two arguments the first is the directory to backup into enstore.
The second is the pnfs directory to place the backups.  All pnfs tags
are left unmodifed and leave the values to the default the parent directories
are set to.  Hardlinks are indistinguishable from each other, thus for
each link a seperate copy is made.  Symbolic links may be transferd... if
the original link points to a file under the same mount point ensync should
be able to create it.

Ensync does one file transfer at a time.  It does not do parallel transfers.
It is not intended to be used for frequent transfers of large numbers of
files.