On a multi-file read, the tape with the last set of files would 'win' in determining the exit status of the entire encp process.  This has been fixed. This did not effect single read requests, multifile reads involving a single tape or writes.
If the routing socket udp_server object failed to reaquire the same port
when reinitializing; the code used to error out.  Now on this 'error' it will
reaquire an unused port.  This may cause some 'Connection timed out' errors
from the mover, but that is life.

Lastly, some underpinnings for printing the CRC in the data_access_layer
have been included.  However, this information is not used yet for anything.