Added carbage collection code.  This is to force collection at otherwise sleeping times while waiting for responces.
Modified handle_retries to be able to check the local file (for existance,
etc.).  This is to be able to better show that encp does not randomly
delete files.

Modified how final_say and delete_at_exit.quit() work together.  This is
to better support the automigration project.

Removed an unceccesary time.sleep() call.

Commented out some debugging log messages.

Tweeked the intervals that the control_socket gets created.