Fixed the bug where entv moves upward on the screen when it reinitializes. The update() function needed to be called for the root window object, not just the Canvas object.
Hopefully, fixed the IRIX issue where the choosen font size is way to
big.  The font family requested was always arial.  The problem is that
few UNIX systems have arial fonts on them.  Linux seemed to pick a good
substitute but IRIX did/does not.  Now uses Helvetica font family.

Added a startup_lock.  This should allow for a faster startup.  This thread
prevents the "netowrk" threads in entv.py from doing anything until the
main thread is in mainloop().  When multiple enstore systems are specified,
it also keeps the resources down by letting only one go at a time.

There was a sizable resouce leak.  The csc's that were created in the
csc_command() function where not being reclaimed.  Despite what gc.collect()
was saying.  These cscs are now explicitly destroyed in destroy() to fix
the problem.

A ton of speed improvements have been added.  Too many to name.  Most are
of types described in Chapter 17 of "Python and Tkinter Programming."
A lot of profiling was done to aid in finding the slowest sections.