Modified how mmap io is handled; this removed an unecessary memory to memory copy.  This also means that for mmap io there is no longer a 'threaded' implimentation (it reverts to single threaded) when both directions of the transfer are using mmap io.  For mmap io to non-mmap io transfers the opposing read()/write() is done directly into the mmap io area; this saves a memory to memory copy.
For ddencp added some new transfer modifiers.  The 0_SYNC, O_DSYNC
and O_RSYNC options can now be added with -S, -D and -R respectively.

Added the -c option.  This tells ddencp to try and clear the file
from the buffer cache.  So far it only seems to work on IRIX.

Added the ability to report how many pages are currently in core
before the start of the transfer.  This is only available on systems
with the mincore() function.  This output is turned on with the verbose
switch (-v).