Fixed a problem at the beginning to recognize when a file is a pnfs or local file w.r.t the check for all three types of paths.  The old way had problems with automounting paths not beginning with /pnfs/...
Found a few more KEYERROR vs. NO_VOLUME/NO_FILE bugs.

The copy numberi, when --copies is used, is relative to this encp.  It is
not absolute w.r.t.  the number of copies of the file in the entire system.
Also, the original bfid and original file family are not mangled on the
encp side.  They will be mangled on the mover side.