Added original_bfid to the file clerk subticket.  (It should have already been there).
Place the copies in the master list of write requests.  Before they were
stored inside the ticket of the original request.  This change was motivated
when printing the requests for the original request were being cut off.

Fixed a bug that compared the conanical path with a hard coded first 19
characters of the path.  19 works for /pnfs/fnal.gov/usr/ but not in the
general case.  Now it uses len(conanical_path).