These modifications are suggestions from Sasha.  Add comments for LOCAL_DIR and PNFS_DIR, REMORTE_ENCP, encp_cmd and debug.  Added comment that get_file_list() and get_pnfs_file_list() initially only consider files in the specified directory; only if no files were found check for files recursively.
One case in cleanup_files() was resulting in rm asking for confirmation due
to missing -f.  Modified volume_assert() and ensync_transfer() to have the
same -d and "-d -d" output that transfer() has.  Patched the pnfs_agent
test to use REMOTE_ENCP only for that test; unless the user specified it
for all tests; previously it was added and never removed for subsequent tests.
Corrected the debug output for some test functions; they had '$encp' instead
of "$encp".