http://uqbar/reviews/r/355/ http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=1040 (mitigates, does not solve)
operation.py currently cuts single ticket with maximum of 5 caps of tapes. Next ticket will be cut of there are more then 10 caps worth of full tapes or 7 days elapsed since last ticket was cut. The volumes in the list are ordered by label. As the result some volume may take a long time to get flipped if they happen to be in the end of the list. The solution is to cut tickets for all tapes in the list. SSA does not want to have one ticket with more than 5 caps. Cutting many tickets, 5 caps each requires some work. Meanwhile a mitigating solution is to order volumes by si_time_1 in ascending order.