http://uqbar/reviews/r/381/ I have conducted a study of stkensrv0n behavior during volume_clerk timeouts:
https://srm.fnal.gov/twiki/bin/view/EnstoreCorner/VolumeClerkTimeouts

Aapparently during accounting db backup activity we see system swapping and queries on enstoredb slowing.
In happens that in the period between midnight and 3:00 AM the activity on enstoredb is particularly heavy. Running accounting db backup during this time compounds the problem.
This patch moves the accounting db backup to 5:10 AM time slot.