New media loader classes and modifications for TS4500 tape library.
This change is to interface media changer and movers with TS4500 tape library using modified mtx tool. The original mtx tool is a command-line interface. The modified mtx tool provides a library of routines to load / unload media into tape drives. It is more effective than the old approach, because it queries robot only once which saves tens of seconds for actual mount / dismount commands. In addition it provides information about tape drive and storage elements and their physical location.

http://srm.fnal.gov/reviews/r/510
http://www-enstore.fnal.gov/Bugzilla/show_bug.cgi?id=1934