monitored_server: remove references to mover and migrator clients.
Inquisitor holds a list of monitored_servers.
Each monitored server corresponds to a Enstore server
(mover, LM, VC ....).

Each monitored server holds on to a client of the
corresponding server. There seems to be a socket associated w/
each client. These are never garbage collected as inqusiotor holds
on to monitored_server references. As number of servers
increase so the number of file descriptors held by inquisitor.

UNKNOWN state observed for the movers was result of
handling "Too many open file descriptors" - inqusitor does
not crash, but it could not update server status also.

This atch removes references to clients from
monitored_server.MonitoredMover and monitored_server.MonitoredMigrator

(monitored_server.MonitoredLibraryManager remains untouched
to minimize changes to inquisitor that uses these classes.
Since # of LMs is small, this should be a good patch for the time being)

	RB: http://srm.fnal.gov/reviews/r/525/
	BZ: http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=1948