Improve mtx command timeout processing.
It was observed that even the mount / dismount command timeouts it actually get executed.
Request database inventory to check if the tape operation was done and retry if it was not.
The database inventory adds about 40s, but it is more reliable to just check if tape was mounted / dismounted rather, than repeat command.
Repetition of the completed command may confise TS4500 firmware as recently observed.
http://srm.fnal.gov/reviews/r/584/
http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=2052