mover: include media changer device into status command return.
Include the media change device name into the status retuned by mover.
This will simplify finding the media changer device for a given mover.
While for SL8500 media change device is a must in the mover configuration,
for TS4500 it is defined, based on 'device' entry in the mover configuration.

http://srm.fnal.gov/reviews/r/615/
http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=2081