SFA migrator. Change (optimize) library tag selection.
Take library tag from the first file in the package.
Curently library tag gets from each file direcroty. The directory is stored in the list and if the directory for the next file is different the library tag get read again and so on.
It was recently noticed that some packages contain files with different directories, which makes selection of the library path extremely inefficient.
It is hard to imagine that files in the same package have different library tags. This is why it is feasible to take the library tag from the first file in the package.
This fix is quick and could be done better, but we currently have no time for making it better.

http://srm.fnal.gov/reviews/r/626/
http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=2097