Mitigate thread switching effect.
It looks as this problem is related to python GLI, which started exposing after kernel upgrade to 2.6.32-754.25.1.el6.x86_64. Tape thread sets state HAVE_BOUND and exits. The main thread detects that tape thread is still running 5 seconds later and generates alarm: "Tape thread is running in the state HAVE_BOUND. Will offline the mover". This was observed only with disk mover and actually mover does not get offlined.
The code was modified to record the occurrence of such condition and wait 1 second.

http://srm.fnal.gov/reviews/r/641/
http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=2117