Modify volume_assert to not stop on errors.
When running volume_assert on FL0016L8 recovery tape, it periodically gets an FTT_EBUSY POSITIONING error which stops the assert.  If at all possibly please modify it so it does not stop when this happens.  IBM does not see any errors (at least for the one I got dump from drive when it occurred for location 29.

http://srm.fnal.gov/reviews/r/669/
http://www-ccf.fnal.gov/Bugzilla/show_bug.cgi?id=2325