Do not remove from at_movers when mover is in SETUP with mounted volume.
Mover is in SETUP state when it tries to connect to client (encp). If mover is in this state for a long time when it has mounted volume, the library manager would remove it from at_mover list. It is free to get mounted in another mover, which causes mount error. This patch fixes such problem.

https://srm.fnal.gov/reviews/r/678/
https://srm.fnal.gov/reviews/r/678/bugs/2066/