Library manager does not delete assigned volume if mover fails to connect to client.
There are 2 changes. One fixes the stated problem.
The second change is to put mover request, coming from mover with bound volumes into postponed bound requests. If there is mover request processing thread running and serve postponed request in the current thread. Before this change the library manager was sending 'nowork' reply to mover in a hope that it would retry later and this retry would be processed ny library manager. This could and would result in expiration of dismount timeout at mover and unnecessary dismount of tape for which there was a potential request.