Fix client behavior for dual stack host configurations.
The previous implementation was allowing to reinit if application
was running on dual stack host and sending to IPV4 server.
This was noticed in encp application. The problem was that retires
were sent on differnt ports, which violates enstores communutaion protocol,
according to which retries must be sent on the same source port.