Production (#9)
* Add sender to log ticket for tcp / ip log client.
Otherwise log server may identify sender from the last udp message which is not always ftrom the same client.

https://www-enstore.fnal.gov/Bugzilla/show_bug.cgi?id=2483

* Library manager does not delete assigned volume if mover fails to connect to client.

There are 2 changes. One fixes the stated problem.
The second change is to put mover request, coming from mover with bound volumes into postponed bound requests. If there is mover request processing thread running and serve postponed request in the current thread. Before this change the library manager was sending 'nowork' reply to mover in a hope that it would retry later and this retry would be processed ny library manager. This could and would result in expiration of dismount timeout at mover and unnecessary dismount of tape for which there was a potential request.

* Added debug output.

* Fix client behavior for dual stack host configurations.

The previous implementation was allowing to reinit if application
was running on dual stack host and sending to IPV4 server.
This was noticed in encp application. The problem was that retires
were sent on differnt ports, which violates enstores communutaion protocol,
according to which retries must be sent on the same source port.

* Removed wrong comments.

* Commented code that may be needed for debugging.

* Get information from movers, associated with Spectra Logic.

These movers do not have 'mc_device' in configuration, just like TS4500 movers.

* Improve security of message parsing. (#7)

Use ast.literal_eval istead of eval.

* ENSTORE_V_6.3.4-12 rpm (#8)

* ENSTORE_V_6.3.4-14