Mc new class (#38)
* Make a new class for Spectra Logic tape library.

In the current implementation the MTXN_MediaLoader class works for both IBM and
Spectra Logic tape libraries. It was noticed that the mount / dismount response
time skipped up noticeably for IBM tape library comparing to the previous
implementation of MTXN_MediaLoader class. This patch copies current MTXN_MediaLoader to MTXN_MediaLoaderSL and reverts MTXN_MediaLoader to its previous implementation.
Testing was done on production system and showed that mount / dismount response
time skipped down back to what is was before.

https://www-enstore.fnal.gov/Bugzilla/show_bug.cgi?id=2534
https://www-enstore.fnal.gov/Bugzilla/show_bug.cgi?id=2530

* Untabified and fixed indentaion where needed in all MTXN classes.

* Changes per PR comments.

* Fixed MTXN_CDB_MediaLoader.getVolState()